Refactor: Feature-Sliced Design 구조 변경 [VirtualFitting_System #73]
