+++ virtualFittingApp/src/pages/order-list/api/order.action.ts
@@ -1,0 +1,6 @@
+import { API_BASILIUM } from "@/shared/config/AxiosConfig";
+
+
+export const fetchOrderHistory = () => {
+    return API_BASILIUM.get("/normalUser/order/history");
+  };
