+++ virtualFittingApp/src/pages/order-list/types/order.d.ts
@@ -1,0 +1,4 @@
+export interface OrderData {
+    orderId: string;
+    orderDate: string;
+  }
