+++ virtualFittingApp/src/shared/components/product/product-order-list/Product_Order_List.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import { ProductOrderListProps } from "../product-order-list/types/product";
 import { formatDate } from "./utils/date.util";
 // import product from '../assets/img/product.svg'; [수정]
 import "./Product_Order_List.css";
